+++ engine/flink/avro-components-utils/src/test/scala/pl/touk/nussknacker/engine/avro/source/flink/TestFromFileSpec.scala
@@ -4,6 +4,7 @@ import com.typesafe.config.ConfigFactory
 import com.typesafe.config.ConfigValueFactory.fromAnyRef
 import com.typesafe.scalalogging.LazyLogging
 import io.circe.Json._
+import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import org.apache.kafka.common.record.TimestampType
 import org.scalatest.{FunSuite, Matchers}
 import pl.touk.nussknacker.engine.api.test.TestData