+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/SchemaBasedMessagesSerdeProvider.scala
@@ -5,13 +5,13 @@ import io.confluent.kafka.schemaregistry.ParsedSchema
 import pl.touk.nussknacker.engine.avro.serialization.{KafkaAvroDeserializationSchemaFactory, KafkaAvroSerializationSchemaFactory}
 import pl.touk.nussknacker.engine.kafka.RecordFormatterFactory
 
-trait SchemaBasedMessagesSerdeProvider extends Serializable {
+trait SchemaBasedMessagesSerdeProvider[T<:ParsedSchema] extends Serializable {
 
-  def deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory
+  def deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory[T]
 
-  def serializationSchemaFactory: KafkaAvroSerializationSchemaFactory
+  def serializationSchemaFactory: KafkaAvroSerializationSchemaFactory[T]
 
   def recordFormatterFactory: RecordFormatterFactory
 
-  def validateSchema[T <: ParsedSchema](schema: T): ValidatedNel[SchemaRegistryError, T]
+  def validateSchema(schema: T): ValidatedNel[SchemaRegistryError, T]
 }