+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/SchemaWithMetadata.scala
@@ -1,8 +1,6 @@
 package pl.touk.nussknacker.engine.avro.schemaregistry
 
 import io.confluent.kafka.schemaregistry.ParsedSchema
-import io.confluent.kafka.schemaregistry.avro.AvroSchema
-import io.confluent.kafka.schemaregistry.client.SchemaMetadata
 
 /**
  * This class holds information that are stored next to schema in registry.
@@ -10,12 +8,3 @@ import io.confluent.kafka.schemaregistry.client.SchemaMetadata
  * is not coupled with concrete schema registry implementation.
  */
 case class SchemaWithMetadata(schema: ParsedSchema, id: Int)
-
-object SchemaWithMetadata{
-  def apply(schemaMetadata: SchemaMetadata): SchemaWithMetadata = {
-    schemaMetadata.getSchemaType match {
-      case "AVRO" => SchemaWithMetadata(new AvroSchema(schemaMetadata.getSchema), schemaMetadata.getId)
-      case other => throw new IllegalArgumentException(s"Not supported schema type: $other")
-    }
-  }
-}