+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/ConfluentAvroSchemaBasedMessagesSerdeProvider.scala
@@ -2,47 +2,38 @@ package pl.touk.nussknacker.engine.avro.schemaregistry.confluent
 
 import cats.data.Validated.{Invalid, Valid}
 import cats.data.{NonEmptyList, ValidatedNel}
-import io.confluent.kafka.schemaregistry.ParsedSchema
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import org.apache.avro.Schema
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.{CachedConfluentSchemaRegistryClientFactory, ConfluentSchemaRegistryClientFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.formatter.{ConfluentAvroToJsonFormatterFactory, JsonPayloadToJsonFormatterFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.jsonpayload.{ConfluentJsonPayloadSerializerFactory, ConfluentKeyValueKafkaJsonDeserializerFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.{ConfluentAvroSerializationSchemaFactory, ConfluentKeyValueKafkaAvroDeserializationFactory}
-import pl.touk.nussknacker.engine.avro.schemaregistry.{SchemaRegistryError, SchemaBasedMessagesSerdeProvider, SchemaRegistryUnsupportedTypeError}
+import pl.touk.nussknacker.engine.avro.schemaregistry.{SchemaBasedMessagesSerdeProvider, SchemaRegistryError, SchemaRegistryUnsupportedTypeError}
 import pl.touk.nussknacker.engine.avro.serialization.{KafkaAvroDeserializationSchemaFactory, KafkaAvroSerializationSchemaFactory}
 import pl.touk.nussknacker.engine.kafka.RecordFormatterFactory
 
 class ConfluentAvroSchemaBasedMessagesSerdeProvider(val schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
-                                                    val serializationSchemaFactory: KafkaAvroSerializationSchemaFactory,
-                                                    val deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory,
-                                                    val recordFormatterFactory: RecordFormatterFactory) extends SchemaBasedMessagesSerdeProvider {
-  override def validateSchema[T <: ParsedSchema](schema: T): ValidatedNel[SchemaRegistryError, T] = {
-    schema match {
-      case s: AvroSchema => {
-        /* kafka-avro-serializer does not support Array at top level
-        [https://github.com/confluentinc/schema-registry/issues/1298] */
-        if (s.rawSchema().getType == Schema.Type.ARRAY)
-          Invalid(NonEmptyList.of(
-            SchemaRegistryUnsupportedTypeError("Unsupported Avro type. Top level Arrays are not supported")))
-        else
-          Valid(schema)
-      }
-      case _ => throw new IllegalArgumentException("Unsupported schema type")
-    }
+                                                    val serializationSchemaFactory: KafkaAvroSerializationSchemaFactory[AvroSchema],
+                                                    val deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory[AvroSchema],
+                                                    val recordFormatterFactory: RecordFormatterFactory) extends SchemaBasedMessagesSerdeProvider[AvroSchema] {
+  override def validateSchema(schema: AvroSchema): ValidatedNel[SchemaRegistryError, AvroSchema] = {
+    /* kafka-avro-serializer does not support Array at top level
+    [https://github.com/confluentinc/schema-registry/issues/1298] */
+    if (schema.rawSchema().getType == Schema.Type.ARRAY)
+      Invalid(NonEmptyList.of(
+        SchemaRegistryUnsupportedTypeError("Unsupported Avro type. Top level Arrays are not supported")))
+    else
+      Valid(schema)
   }
 }
 
 object ConfluentAvroSchemaBasedMessagesSerdeProvider extends Serializable {
 
-  def apply(): ConfluentAvroSchemaBasedMessagesSerdeProvider =
-    avroPayload(CachedConfluentSchemaRegistryClientFactory)
-
   def apply(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
-    avroPayload(schemaRegistryClientFactory)
+    avroSchemaAvroPayload(schemaRegistryClientFactory)
   }
 
-  def avroPayload(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
+  def avroSchemaAvroPayload(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
     ConfluentAvroSchemaBasedMessagesSerdeProvider(
       schemaRegistryClientFactory,
       new ConfluentAvroSerializationSchemaFactory(schemaRegistryClientFactory),
@@ -51,19 +42,7 @@ object ConfluentAvroSchemaBasedMessagesSerdeProvider extends Serializable {
     )
   }
 
-  def apply(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
-            serializationSchemaFactory: KafkaAvroSerializationSchemaFactory,
-            deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory,
-            recordFormatterFactory: RecordFormatterFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
-    new ConfluentAvroSchemaBasedMessagesSerdeProvider(
-      schemaRegistryClientFactory,
-      serializationSchemaFactory,
-      deserializationSchemaFactory,
-      recordFormatterFactory
-    )
-  }
-
-  def jsonPayload(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
+  def avroSchemaJsonPayload(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
     ConfluentAvroSchemaBasedMessagesSerdeProvider(
       schemaRegistryClientFactory,
       new ConfluentJsonPayloadSerializerFactory(schemaRegistryClientFactory),
@@ -72,4 +51,15 @@ object ConfluentAvroSchemaBasedMessagesSerdeProvider extends Serializable {
     )
   }
 
+  private def apply(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
+                    serializationSchemaFactory: KafkaAvroSerializationSchemaFactory[AvroSchema],
+                    deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory[AvroSchema],
+                    recordFormatterFactory: RecordFormatterFactory): ConfluentAvroSchemaBasedMessagesSerdeProvider = {
+    new ConfluentAvroSchemaBasedMessagesSerdeProvider(
+      schemaRegistryClientFactory,
+      serializationSchemaFactory,
+      deserializationSchemaFactory,
+      recordFormatterFactory
+    )
+  }
 }