+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/serialization/KafkaAvroSerializationSchemaFactory.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.engine.avro.serialization
 
+import io.confluent.kafka.schemaregistry.ParsedSchema
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import org.apache.flink.formats.avro.typeutils.NkSerializableAvroSchema
 import pl.touk.nussknacker.engine.kafka.{KafkaConfig, serialization}
@@ -10,8 +11,8 @@ import pl.touk.nussknacker.engine.util.KeyedValue
   * of KafkaSerializationSchema are needed additional avro related information. SerializationSchema will take
   * KafkaSerializationSchema with key extracted in the step before serialization
   */
-trait KafkaAvroSerializationSchemaFactory extends Serializable {
+trait KafkaAvroSerializationSchemaFactory[T<:ParsedSchema] extends Serializable {
 
-  def create(topic: String, version: Option[Int], schemaOpt: Option[NkSerializableAvroSchema[AvroSchema]], kafkaConfig: KafkaConfig): serialization.KafkaSerializationSchema[KeyedValue[AnyRef, AnyRef]]
+  def create(topic: String, version: Option[Int], schemaOpt: Option[NkSerializableAvroSchema[T]], kafkaConfig: KafkaConfig): serialization.KafkaSerializationSchema[KeyedValue[AnyRef, AnyRef]]
 
 }