+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/sink/KafkaAvroSinkFactory.scala
@@ -30,7 +30,7 @@ object KafkaAvroSinkFactory {
 }
 
 class KafkaAvroSinkFactory(val schemaRegistryClientFactory: SchemaRegistryClientFactory,
-                           val schemaBasedMessagesSerdeProvider: SchemaBasedMessagesSerdeProvider,
+                           val schemaBasedMessagesSerdeProvider: SchemaBasedMessagesSerdeProvider[AvroSchema],
                            val processObjectDependencies: ProcessObjectDependencies,
                            implProvider: KafkaAvroSinkImplFactory)
   extends KafkaAvroBaseTransformer[Sink] with SinkFactory {