+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/source/SpecificRecordKafkaAvroSourceFactory.scala
@@ -1,6 +1,7 @@
 package pl.touk.nussknacker.engine.avro.source
 
 import cats.data.Validated.Valid
+import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import org.apache.avro.specific.SpecificRecord
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.context.transformation.{DefinedEagerParameter, NodeDependencyValue}
@@ -17,7 +18,7 @@ import scala.reflect.{ClassTag, classTag}
  * Source factory for specific records - mainly generated from schema.
  */
 class SpecificRecordKafkaAvroSourceFactory[V <: SpecificRecord: ClassTag](schemaRegistryClientFactory: SchemaRegistryClientFactory,
-                                                                          schemaBasedMessagesSerdeProvider: SchemaBasedMessagesSerdeProvider,
+                                                                          schemaBasedMessagesSerdeProvider: SchemaBasedMessagesSerdeProvider[AvroSchema],
                                                                           processObjectDependencies: ProcessObjectDependencies,
                                                                           implProvider: KafkaSourceImplFactory[Any, V])
   extends KafkaAvroSourceFactory[Any, V](schemaRegistryClientFactory, schemaBasedMessagesSerdeProvider, processObjectDependencies, implProvider) {